2048 fonctionne presque mais vraiment presque

+++ implem/ei_implementation.h
@@ -119,10 +119,10 @@ typedef struct ei_impl_placer_params_t
 	int y;
 	int width;
 	int height;
-	int rel_x;
-	int rel_y;
-	int rel_width;
-	int rel_height;
+	float rel_x;
+	float rel_y;
+	float rel_width;
+	float rel_height;
 
 } ei_impl_placer_params_t;
